inherit ignore unless overriden in the derived map; fixes #1910; fixes #1732
